pyproject.toml
[project]
name = "microgrid-rl-env"
version = "1.0.0"
description = "Microgrid RL OpenEnv environment"
readme = "README.md"
requires-python = ">=3.10"

authors = [
    {name = "Anirudh Patil"}
]

dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
